OrcLib: Archive: fix incorrect use of size_t with uint64_t
This fix an issue with 7z which would refuse to work with 32 bit sizes.
Only GetThis was affected.

+++ src/OrcLib/Archive/Item.cpp
@@ -32,7 +32,7 @@ std::shared_ptr<ByteStream>& Item::Stream()
     return m_stream;
 }
 
-size_t Item::Size() const
+uint64_t Item::Size() const
 {
     if (m_stream == nullptr)
     {
